+++ src/editor_helpers.ts
@@ -1,5 +1,5 @@
 import { Platform } from "obsidian";
-import { Environment } from "./snippets";
+import { Environment } from "./snippets/snippets";
 import { EditorView } from "@codemirror/view";
 import { EditorSelection, SelectionRange, EditorState } from "@codemirror/state";
 import { syntaxTree } from "@codemirror/language";
